Renamed stack files.
