+++ src/objects/alloc_stack.c
@@ -1,4 +1,4 @@
-#include "function_stack.h"
+#include "alloc_stack.h"
 
 void *s_alloc(stack *stack, size_t n) {
   //fprintf(stdout, "Requested %ld bytes from buffer of size %ld.\n", n, stack->size);
