+++ src/objects/alloc_stack.h
